database created

+++ db/migrate/20101022113803_create_products.rb
@@ -2,7 +2,7 @@ class CreateProducts < ActiveRecord::Migration
   def self.up
     create_table :products do |t|
       t.string :name
-      t.double :price
+      t.decimal :price
 
       t.timestamps
     end
